# API-Gateway/openapi.yaml
openapi: 3.1.0
info:
  title: FastAPI
  version: 0.1.0
paths:
  /api/sign-up:
    post:
      summary: Sign Up
      operationId: sign_up_api_sign_up_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserSignUpRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/sign-in:
    post:
      summary: Sign In
      operationId: sign_in_api_sign_in_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserSignInRequest'  # ❗ ИСПРАВЛЯЕМ: SignIp -> SignIn
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSignInResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  
  # ❗ ДОБАВЛЯЕМ НОВЫЕ ЭНДПОИНТЫ ДЛЯ TELEGRAM
  /api/link_telegram:
    post:
      summary: Link Telegram Account
      operationId: link_telegram_api_link_telegram_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TelegramLinkRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  
  /api/auth/telegram:
    post:
      summary: Auth by Telegram ID
      operationId: auth_telegram_api_auth_telegram_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TelegramAuthRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  
  /api/user/by-telegram/{telegram_id}:
    get:
      summary: Get User by Telegram ID
      operationId: get_user_by_telegram_api_user_by_telegram__telegram_id__get
      parameters:
        - name: telegram_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TelegramUserResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'

  /api/predict:
    post:
      summary: Predict
      operationId: predict_api_predict_post
      security:
        - OAuth2PasswordBearer: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AIRequest'
        required: true
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'

components:
  schemas:
    # ❗ ДОБАВЛЯЕМ НОВЫЕ МОДЕЛИ ДЛЯ TELEGRAM
    TelegramAuthRequest:
      title: TelegramAuthRequest
      required:
        - telegram_id
      type: object
      properties:
        telegram_id:
          title: Telegram Id
          type: string
    
    TelegramLinkRequest:
      title: TelegramLinkRequest
      required:
        - email
        - password
        - telegram_id
      type: object
      properties:
        email:
          title: Email
          type: string
        password:
          title: Password
          type: string
        telegram_id:
          title: Telegram Id
          type: string
        telegram_username:
          title: Telegram Username
          type: string
    
    TelegramUserResponse:
      title: TelegramUserResponse
      required:
        - user_id
        - username
      type: object
      properties:
        user_id:
          title: User Id
          type: integer
        username:
          title: Username
          type: string
        email:
          title: Email
          type: string
        telegram_id:
          title: Telegram Id
          type: string
        telegram_username:
          title: Telegram Username
          type: string
        first_name:
          title: First Name
          type: string
        last_name:
          title: Last Name
          type: string
    
    AIRequest:
      title: AIRequest
      required:
        - category
        - time
        - cords
        - place
      type: object
      properties:
        category:
          title: Category
          type: array
          items: {}
        time:
          title: Time
          type: number
        cords:
          title: Cords
          type: string
        place:
          title: Place
          type: string
    
    AcceptResponse:
      title: AcceptResponse
      required:
        - user_id
        - task_id
      type: object
      properties:
        user_id:
          title: User Id
          type: integer
        task_id:
          title: Task Id
          type: string
    
    HTTPValidationError:
      title: HTTPValidationError
      type: object
      properties:
        detail:
          title: Detail
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
    
    UserSignInResponse:
      title: UserSignInResponse
      required:
        - token
        - user_id
        - username
        - message
      type: object
      properties:
        token:
          title: Token
          type: string
        user_id:
          title: User Id
          type: integer
        username:
          title: Username
          type: string
        message:
          title: Message
          type: string
    
    UserSignInRequest:  # ❗ ИСПРАВЛЯЕМ: SignIp -> SignIn
      title: UserSignInRequest
      required:
        - email
        - password
      type: object
      properties:
        email:
          title: Email
          type: string
        password:
          title: Password
          type: string
    
    UserSignUpRequest:
      title: UserSignUpRequest
      required:
        - username
        - email
        - password
      type: object
      properties:
        username:
          title: Username
          type: string
        email:
          title: Email
          type: string
        password:
          title: Password
          type: string
    
    ValidationError:
      title: ValidationError
      required:
        - loc
        - msg
        - type
      type: object
      properties:
        loc:
          title: Location
          type: array
          items:
            anyOf:
              - type: string
              - type: integer
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
  
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          tokenUrl: /api/sign-in
          scopes: {}